Custom CRSC CUDA kernel pipeline, GPU energy profiling, and efficiency gap metric
- Implemented LIF CUDA kernel (src/crsc/kernels/) with normal and energy-profiled forward pass via PyBind11
- Added GPU attribute modules: EnergyProfiler (NVML power + CUDA event timing), memory management, occupancy-optimal launch config
- Flattened acceleration/N_H/ into acceleration/ and moved kernels from src/learning/ to src/crsc/
- Gated custom kernel dispatch behind kernel: ON/OFF in SNN_module.yaml across trainer and tester
- Added NVML power sampling to GPUStats; inference summary now reports actual GPU energy alongside the neuromorphic pJ estimate and prints the hardware efficiency gap ratio
- Fixed adversarial robustness crash caused by int32 targets passed to cross_entropy)

+++ src/learning/adversarial_robustness.py
@@ -85,8 +85,8 @@ class AdversarialEvaluator:
         correct = total = 0
 
         for data, targets in self.test_loader:
-            data = data.to(self.device)
-            targets = targets.to(self.device)
+            data    = data.to(self.device)
+            targets = targets.to(self.device).long()
 
             if attack == "fgsm":
                 data = generate_fgsm_input(self.model, data, targets, epsilon)
